% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/crop_function.R
\name{crop_function}
\alias{crop_function}
\title{A Cat Function}
\usage{
crop_function(love = TRUE)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

